move buttons left
