+++ pkg/services/featuremgmt/registry.go
@@ -148,12 +148,6 @@ var (
 			State:       FeatureStateStable,
 		},
 		{
-			Name:         "saveWithDiff",
-			Description:  "Open a drawer view with diff for save modal",
-			State:        FeatureStateAlpha,
-			FrontendOnly: true,
-		},
-		{
 			Name:        "dashboardComments",
 			Description: "Enable dashboard-wide comments",
 			State:       FeatureStateAlpha,
